backend/requirements.txt
fastapi
uvicorn
gunicorn
psycopg2-binary
python-dotenv
pydantic
sqlalchemy
passlib
bcrypt
python-jose
